Reimplement mirrors arithmetically, replacing the lookup table.
Many thanks to @yaahc for figuring out the math(s) here. The idea is to represent mirror angles using the same system as thread directions, and then to calculate the angle of incidence as if it were rotation. Except angles are modulo 8, using the same coordinate system again, so we can rotate with addition.